+++ resources/views/purchases/edit.blade.php
@@ -24,7 +24,7 @@
     @foreach($valid as $k=>$v)
     <?php if (in_array('date',$v)) { ?>
       <x-adminlte-input name="{{ $k; }}" label="{{ $k; }}" type="date" value="{{ $res->{k}; }}"/>
-    <?php } else if (in_array('integer', $v)) {
+    <?php } else if (in_array('numeric', $v)) {
       $step = "0";
       if ($k == 'Price') { $step = "0.01";}
     ?>
